modified:   hunvec/corpus/tagged_corpus.py modified:   hunvec/seqtag/word_tagger.py modified:   hunvec/seqtag/word_tagger_dataset.py modified:   hunvec/utils/data_splitter.py huge amount of change, now first converging version

+++ hunvec/utils/data_splitter.py
@@ -4,9 +4,9 @@ from random import shuffle
 def datasplit(data, ratios=[.7, .15, .15]):
     indices = range(len(data))
     shuffle(indices)
-    training = ratios[0] * len(data)
-    test = ratios[1] * len(data)
-    valid = ratios[2] * len(data)
+    training = int(ratios[0] * len(data))
+    test = int(ratios[1] * len(data))
+    valid = int(ratios[2] * len(data))
     training_indices = indices[:training]
     valid_indices = indices[training:training + valid]
     test_indices = indices[training+valid:training+valid+test]
